Add built-in pipe
